% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{a.function}
\alias{a.function}
\title{Add two numbers}
\usage{
a.function(a, b)
}
\arguments{
\item{a}{The first number to be added}

\item{b}{The second number to be added}
}
\value{
a+b
}
\description{
Add two numbers
}
