% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{give_paths}
\alias{give_paths}
\title{Assign the global paths}
\usage{
give_paths()
}
\description{
Assign \code{.datadir} (owncloud with the data) and \code{.workdir} to be used as data and working directory.
}
